Reconciles the two forked copies of the M1 solver, keeping cme-multifreq's
chiral magnetic effect and largesim-m1's non-thermal separated opacities,
partial-equilibrium predictor and STS/parabolic task lists.

Adopts largesim-m1's UnitSystem convention (numberDensity constant fm^-3,
length scale in the new volume entry). The multi-frequency solver's
nudens_0 is in code units, so its nurates conversion uses
1/VolumeConversion to keep the length factor.

+++ src/outputs/binary.cpp
@@ -97,7 +97,7 @@ void MeshBinaryOutput::WriteOutputFile(Mesh *pm, ParameterInput *pin) {
         << "  size of variable=" << sizeof(float) << std::endl
         << "  number of variables=" << outvars.size()+outvars_6d.size() << std::endl
         << "  variables:  ";
-    for (int n=0; n<outvars.size(); n++) {
+    for (size_t n=0; n<outvars.size(); n++) {
       msg << outvars[n].label.c_str() << "  ";
     }
     if (out_params.n_derived_6d > 0) {
